+++ test/libsolidity/smtCheckerTests/external_calls/external_hash_known_code_state_reentrancy_unsafe.sol
@@ -4,7 +4,7 @@ contract State {
 	uint x;
 	C c;
 	function f() public returns (uint) {
-		c.setOwner(address(0));
+		c.setOwner(address(uint160(0)));
 		return c.g();
 	}
 }
@@ -34,6 +34,6 @@ contract C {
 	}
 }
 // ----
-// Warning 5084: (116-126): Type conversion is not yet fully supported and might yield false positives.
-// Warning 4661: (388-402): Assertion violation happens here
-// Warning 4661: (406-432): Assertion violation happens here
+// Warning 5084: (124-134): Type conversion is not yet fully supported and might yield false positives.
+// Warning 4661: (397-411): Assertion violation happens here
+// Warning 4661: (415-441): Assertion violation happens here
